implemented reporting for posts in feed and profile
