+++ client/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "_proxy": "http://localhost:5000",
   "dependencies": {
+    "@emailjs/browser": "^3.10.0",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@fortawesome/fontawesome-svg-core": "^6.2.0",
